+++ nix/cardano-bridge.nix
@@ -12,7 +12,7 @@ in runCommand "daedalus-cardano-bridge" {
   mkdir -pv $out/bin
   cd $out/bin
   echo ${cardano-wallet.version} > $out/version
-  cp ${cardano-wallet.haskellPackages.cardano-wallet-byron.components.exes.cardano-wallet-byron}/bin/cardano-wallet-byron* .
+  cp ${cardano-wallet.cardano-wallet-byron}/bin/cardano-wallet-byron* .
   cp ${cardano-shell.nix-tools.cexes.cardano-launcher.cardano-launcher}/bin/cardano-launcher* .
   cp ${cardano-node}/bin/cardano-node* .
   cp ${export-wallets}/bin/export-wallets* .